Moved definition of tril/triu into dedicated file
